authorization implemented

+++ src/api/notes/handler.js
@@ -16,8 +16,10 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
- 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+
+      const noteId = await this._service.addNote({ title, body, tags, owner: credentialId,
+      });
  
       const response = h.response({
         status: 'success',
@@ -50,7 +52,8 @@ class NotesHandler {
   }
  
   async getNotesHandler() {
-    const notes = await this._service.getNotes();
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -62,13 +65,17 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
+      
       return {
         status: 'success',
         data: {
           note,
         },
-      };
+       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
@@ -78,7 +85,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+   
       // Server ERROR!
       const response = h.response({
         status: 'error',
@@ -89,15 +96,16 @@ class NotesHandler {
       return response;
     }
   }
+  
  
   async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const { title, body, tags } = request.payload;
       const { id } = request.params;
- 
-      await this._service.editNoteById(id, { title, body, tags });
- 
+      const { id: credentialId} = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
+      await this._service.editNoteById(id, request.payload);
+   
       return {
         status: 'success',
         message: 'Catatan berhasil diperbarui',
@@ -111,7 +119,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+   
       // Server ERROR!
       const response = h.response({
         status: 'error',
@@ -121,13 +129,15 @@ class NotesHandler {
       console.error(error);
       return response;
     }
-  }
+  }  
  
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
- 
+   
       return {
         status: 'success',
         message: 'Catatan berhasil dihapus',
@@ -141,7 +151,7 @@ class NotesHandler {
         response.code(error.statusCode);
         return response;
       }
- 
+   
       // Server ERROR!
       const response = h.response({
         status: 'error',
@@ -152,6 +162,7 @@ class NotesHandler {
       return response;
     }
   }
+  
 }
  
 module.exports = NotesHandler;
